enable more ci for main branch (#86)
* enable more ci for main branch

* remove unused header files

* fix cosmos type issue

use string type instead of int in case of cosmos-like chain value more than int64

* add iost HDWallet address unittest

* keep proto header files

* fix eos unittest issue

* fix conversion warning

* fix some conversion warning

* fix cosmos type issue

* fix swift eos signing issue

* fix swift unittest issue

+++ tests/interface/TWTransactionCompilerTests.cpp
@@ -653,7 +653,7 @@ TEST(TWTransactionCompiler, ExternalSignatureSignRipple) {
     auto preImageHashData = dataFromTWData(preImageHashTWData);
     auto preSigningOutput = TW::TxCompiler::Proto::PreSigningOutput();
     ASSERT_TRUE(
-        preSigningOutput.ParseFromArray(preImageHashData->data(), preImageHashData->size()));
+        preSigningOutput.ParseFromArray(preImageHashData->data(), (int)preImageHashData->size()));
     ASSERT_EQ(preSigningOutput.errorcode(), 0);
     // preSigningOutput.ParseFromArray(preImageHashData.data(), (int)preImageHashData.size());
     auto preImage = preSigningOutput.data();
